sdk/python/pyproject.toml
[build-system]
requires = ["poetry>=1.0.9"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "approzium"
version = "0.2.0"
description = "Approzium SDK provides Approzium database authentiation for Python"
authors = ["Cyral <security@cyral.org>"]
homepage = "https://github.com/cyralinc/approzium"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.5"
boto3 = "^1.14.10"
grpcio = "^1.30.0"
grpcio-tools = "^1.30.0"
asyncpg = { version = "^0.20.1", optional = true}
mysql-connector-python = { version = "^8.0.20", optional = true}
opentelemetry-ext-psycopg2 = { version = "^0.10b0", optional = true}
opentelemetry-ext-jaeger = {version = "^0.10b0", optional = true}
ec2-metadata = { version = "^2.2.0", optional = true}
psycopg2 = { version = "^2.8.5", optional = true}
pymysql = {version = "^0.9.3", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytest-parallel = "^0.1.0"
pytest-asyncio = "^0.14.0"

[tool.poetry.extras]
sqllibs = ["psycopg2", "asyncpg", "mysql-connector-python", "pymysql"]
tracing = ["opentelemetry-ext-psycopg2", "opentelemetry-ext-jaeger", "ec2-metadata"]

[tool.black]
target-version = ['py37']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.hg          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | approzium/_protos
)
'''
